+++ adapter/bidmachine/CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
-- Relayed ML floor predictions from the bid response custom params into the ad unit ext reported in stats and auction info
+- Relayed the bid response custom params to the ad unit ext reported in stats and auction info
 
 ## [3.7.1.0] - 2026-06-01
 ### Changed